backend/app.py
"""
AI-Powered Real Estate Intelligence Platform for Tanzania
Main FastAPI Application
"""

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
from pydantic import BaseModel
from typing import List, Optional
import uvicorn
import os
from dotenv import load_dotenv

# Import routes
from routes.properties import router as properties_router
from routes.predictions import router as predictions_router
from routes.analytics import router as analytics_router

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="Tanzania Real Estate AI",
    description="AI-powered real estate intelligence platform for Tanzania",
    version="1.0.0",
    docs_url="/docs",
    redoc_url="/redoc"
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=[
        "https://tanzania-real-estate-ai-frontend.netlify.app",
        "http://localhost:3000",
        "http://127.0.0.1:3000",
        "http://localhost:5173",
        "http://127.0.0.1:5173"
    ],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Include routers
app.include_router(properties_router, prefix="/api/properties", tags=["properties"])
app.include_router(predictions_router, prefix="/api/predictions", tags=["predictions"])
app.include_router(analytics_router, prefix="/api/analytics", tags=["analytics"])

# Health check endpoint
@app.get("/health", tags=["health"])
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "message": "Tanzania Real Estate AI API is running",
        "version": "1.0.0"
    }

# Root endpoint
@app.get("/", tags=["root"])
async def root():
    """Root endpoint"""
    return {
        "message": "Welcome to Tanzania Real Estate AI API",
        "docs": "/docs",
        "health": "/health"
    }

# Global exception handler
@app.exception_handler(Exception)
async def global_exception_handler(request, exc):
    """Global exception handler"""
    return JSONResponse(
        status_code=500,
        content={
            "error": "Internal Server Error",
            "message": str(exc),
            "type": type(exc).__name__
        }
    )

if __name__ == "__main__":
    uvicorn.run(
        "app:app",
        host="0.0.0.0",
        port=int(os.getenv("PORT", 8000)),
        reload=True
    )
