requirements.txt
# Core ML & Data Science
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
matplotlib==3.7.2
seaborn==0.12.2

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
redis==5.0.1

# Data Collection & Processing
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.3

# ML Model Persistence
joblib==1.3.2

# Data Validation
python-dotenv==1.0.0
pydantic-settings==2.1.0

# API Documentation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Advanced ML (optional)
tensorflow==2.14.0
torch==2.1.1
transformers==4.35.2

# Geospatial Analysis
geopandas==0.14.0
folium==0.15.0

# Time Series Analysis
prophet==1.1.5

# Performance Monitoring
prometheus-client==0.19.0
